sched/ghost: change MSG_TASK_LATCHED to MSG_TASK_ON_CPU
The old MSG_TASK_LATCHED is really TASK_ON_CPU: when you get on cpu.
The actual 'latched notification' is when bpf_ghost_run_gtid() returns
with success (recall that the RQ lock is held now).

Rename TASK_LATCHED to TASK_ON_CPU, and do not send MSG_TASK_LATCHED
before you get on cpu.

There was one corner case where TASK_LATCHED was sent before you got on
cpu: latched_preemption.  There was an old scenario where it was
possible to see the LATCH before the prev task got off cpu, which was
confusing.  That case involved a userspace global agent issuing
transactions for remote cpus, racing with BPF latching.  That's why SGS
used to check msg_latched->latched_preempt.  Not sure if that case is
possible with a purely BPF agent.

Instead of checking latched_preempt, the agent can just check
msg_preempt->was_latched.  And if the agent wants MSG_TASK_ON_CPU, it
knows the task actually got on cpu, so it doesn't need to worry about
weird ordering.

In this sense, you can think of MSG_TASK_PREEMPT as "the ball is yours
for whatever reason" (i.e. the task is runnable), and was_latched tells
you whether it was on cpu or not.  Alternatively, the agent can infer
was_latched based on whether or not it saw TASK_ON_CPU.

Note: I went with TASK_ON_CPU instead of TASK_RUNNING, since in the
kernel, TASK_RUNNING is on_rq, not merely on_cpu.  Confusing, but ON_CPU
leaves no room for doubt.

Change-Id: I8f2b02c56e6a6566f5783af1d296d5b05dcce8cd