sched/ghost: move bpf_cannot_load_prog
sched_ghost_entity_init() gets called both when we fork and when we
are setsched into ghost.  In the latter case, if someone other than us
moves us into ghost, we get their bpf_cannot_load_prog taint, which
could also clobber our taint!

Additionally, bpf_cannot_load_prog isn't actually ghost-class-task
specific.  It's more like inhibit_task_msgs, where it's a property of
the task in general, and not just while it is in the ghost class.  So it
shouldn't even be in struct sched_ghost_entity.

Change-Id: I43bd8d78bc516ca661a0e6464f31956149ec821e