sched/ghost: move __target_enclave
__target_enclave was in p->ghost, but it's used regardless of whether a
task is in the ghost class or not.  It's use is largely as task-local
storage for ghost-bpf to identify which enclave we are operating on
behalf of.

Specifically, we had a race where we could hit the WARN in
bpf_ghost_run_gtid(), where there was no __target_enclave.  During
pnt_prologue, once we unlock the RQ, there's no protection for
current->ghost.__target_enclave.  If another cpu moves that task into
ghost, prior to this commit, we'd zero-clobber the __target_enclave.  If
that happened during BPF-PNT, when we popped out of BPF to run the
helper, the __target_enclave would be zero.  Whoops.

The fix is to move __target_enclave out of struct sched_ghost_entity.
You should only ever modify it via current.

Change-Id: I40fe69dfbe6201b096788844b4dd1d8850691418