sched/ghost: add MSG_TASK_LATCH_FAILURE
When we fail to run a task from BPF-PNT, it helps the BPF agent to find
out about the failure while we hold the task's RQ lock.  The task lock
serializes access to the thread struct in BPF, and lets the agent
clearly know "who has the ball" - specifically between LATCH_FAILURE and
an asynchronous message like DEPARTED.

Since we are failing to latch the task, there's no guarantee that the
task's RQ lock is this cpu's RQ lock.  For instance, in the EBUSY case,
the task might not be migratable from another cpu yet.  The only time
that we hold the task's RQ lock during a failure is when we notice the
failure, e.g. right after out_unlock with an error return value.

The slightly tricky bit is that we need to ensure there will not be two
BPF-MSGs on the same cpu at the same time.  Since ghost_run_gtid() is
only called from BPF-PNT, and BPF-MSG cannot run at the same time as
BPF-PNT, we're OK.

If it ever turns out that we want to run multiple BPF-MSG at the same
time, e.g. trigger some message during a handler like
bpf_ghost_resched_cpu(), we could always move the per-cpu struct
bpf_ghost_msg onto the caller's stack.  The reason for the percpu
variable is to avoid zeroing the entire struct, which you'd need to do
if it was on the stack to avoid data leaks.

Change-Id: Ic693cbca115d0482227542fbb9f3e177f3883126