sched/ghost: always invalidate latched tasks in dequeue
Anytime we dequeue a latched task, regardless of the reason, kick it
back to the agent with a Preempt.

Previously, for 'spurious' events (DEQUEUE_SAVE), we left the task
latched.  DEQUEUE_SAVE is used for stuff like prio_change: something
about the task is changing, and core.c will dequeue, make a change, and
reenqueue.

Currently, for tasks that are on_cpu (not merely latched), we kick them
off the cpu and hand them back to the agent.  By also kicking out
latched tasks, we treat them the same as as if they were running: hand
them back to the agent before changing the prio.  In this way, the agent
has one less state to worry about.

Note that the agent already has to deal with the latched -> preempt ->
prio_change case, which can happen if CFS preempts a latched task.  With
this commit, the agent no longer needs to deal with latched ->
prio_change.

Arguably, this isn't a big deal - the agent knows the task was latched,
but not on_cpu.  And this doesn't help with pending_latch: the agent's
mechanism for dealing with tasks that get dequeued and operated on (e.g.
prio, departed, affinity) before they are latched.

Change-Id: Ic83e3079f0e3f461823eeda58c21eb63176d839e