sched/ghost: set ghost.new_task in ghost_setscheduler()
When a task joins ghost, there is a window of time where it's in the
ghost sched class, we haven't sent a task_new, but we also did not set
ghost.new_task, which is our usual flag for "do we still need to send
MSG_TASK_NEW".

Here's how setscheduler from CFS into ghost worked prior to this commit:

__sched_setscheduler()
- ghost_setscheduler()
	- p->ghost.new_task wasn't set yet
	- this is where this commit sets p->ghost.new_task
- dequeue_task() (CFS)
- put_prev_task() (CFS)
- __set_scheduler()
	- change p->sched_class = &ghost_sched_class,
	- p->ghost.new_task wasn't set yet
- enqueue_task() (Ghost)
- set_curr_task() (Ghost)
	- p->ghost.new_task wasn't set yet
- check_class_changed()
	- switched_to_ghost()
		- send MSG_TASK_NEW or set ghost.new_task

After __set_scheduler(), we're calling enqueue_task and set_curr_task
for Ghost, but new_task doesn't actually tell us if the task is new or
not.  Set ghost.new_task in ghost_setscheduler to cover this gap.

We could have set ghost.new_task in __set_scheduler(), but that's in
core.c, and would require adding more ABI functions.

Change-Id: Ic294fd2a6b513a9c6068530b0a75a7a332eab431