sched: allow ghost to route agent wakeups for messages
When a message is generated, BPF already has the ability to intercept
and do processing based on the message type. This patch extends the
capability of this BPF program type to also allow BPF to select the
preferred cpu to wake up in response to the message (if BPF chooses to
allow the message to be produced into a channel). This will be used for
various performance optimizations.

Change-Id: I538961deaff9f0ec9e703ae046f7fcc418726834