sched/ghost: do not force prio/affinity changes off cpu
The agent gets a message for e.g. a prio change.  If it wants the task
off the cpu, it can issue a transaction (userspace) or a resched_cpu
(bpf).

Note that the kernel is a little inconsistent with the ordering of
operations, depending on the type of event.  For a running task, the
flow is typically:
- dequeue_task()
- put_prev_task()
- do the work
	__setscheduler_prio(),
	p->static_prio = NICE_TO_PRIO(prio),
	p->sched_class->set_cpus_allowed(),
	__setscheduler_uclamp,
	p->numa_preferred_nid = nid,
	sched_change_group(tsk, TASK_MOVE_GROUP),
	etc,
- enqueue_task()
- set_curr_task() (a.k.a. set_next_task()

but for some of these, we run the sched op *after* set_curr_task, e.g.:
- check_class_changed() (which might run prio_changed())
- prio_changed()

So for prio_change, set_curr_task() happens *before* ghost_prio_changed
(and enqueue).  But set_curr_task() happens *after* affinity_change.

It doesn't affect this patch directly, but if you wanted to do something
like "send a preempt message from set_curr_task", that wouldn't give the
agent consistent behavior when it comes to dealing with on_cpu tasks
that have a change to their attributes.

Change-Id: Ie86e02174eeb79d1372825857b4f685ff4e9977c