sched/ghost: return ESTALE for tasks not in ghost
This simplifies the agent's handling of Departed and
bpf_ghost_run_gtid().  EINVAL is reserved for serious errors, worthy at
least of a print/warn.

Tasks that departed from ghost are not in ghost.  Tasks that were
never in ghost are also not in ghost.  We make no distinction, and
arguably there isn't one: consider a task that departed from ghost
weeks ago, should we still say "oh, it was in ghost a long time ago!".

By returning ESTALE, we're saying "your information is out of date".  If
the agent attempts to schedule a task that was never in ghost at all,
arguably that is more of a bug than ESTALE, but I'm less concerned about
agent's manufacturing gtids (trying to run a task the agent *never*
received a TASK_NEW for) than legitimate behavior (trying to run a task
that recently Departed).

Change-Id: Ia3e91867095813c4b28c9bb6d955d1a78f1681f9