sched/ghost: use select_task_rq()'s cpu instead task_cpu
The cpu parameter is what the kernel core wants the task's cpu to be.

select_task_rq() takes cpu equal to either task_cpu() directly or
p->wake_cpu.  p->wake_cpu is set in __set_task_cpu() to the same cpu
that task_cpu() returns (p->cpu).  There are a couple places where the
kernel changes wake_cpu for task migration, where it wants the task's
cpu to be the new cpu.

Change-Id: I7c8ecf4e2ef72bb4528a39a0b93d696d06764225