sched/ghost: use a single bpf func_proto
The func_proto tells the verifier whether or not our program type can
call certain helpers.  Every program type can be attached at only one
place, which is the 'expected attach type', so we can use a single
function to express our helper policy.

Using a single function cuts down on copy-paste issues and potentially
errors.  The main safety concern is that run_gtid is called only from
BPF-PNT.  (There's a check in run_gtid to enforce this).

Note that "ghost_sched" is BPF-PNT.  The name is older, since the
original ghost_sched type was for tick suppression.

Change-Id: I4bb09cc8de0b053c7e0bbd680e77714bd60968ec