sched/ghost: add bpf_select_rq()
Allow the BPF agent to both pick the kernel RQ/cpu that the task wakes
on, as well as choose whether to enable the TTWU_QUEUE mechanism.

If you skip_ttwu_queue, the kernel will grab the (possibly remote) cpu's
RQ and perform the wakeup on the calling cpu.

Note that it is dangerous to use TTWU_QUEUE if you route all wakeups to
the same CPU.  Under heavy load, you could get that cpu stuck endlessly
handling resched IPIs.

Change-Id: Ibb1297802fbb0fd6bbcd5df58ea0aef0b1daeb79