sched: allow ghost messages to elide agent wake
This allows BPF to have a message get sent to the agent, but not
actually wake the agent.

This is useful when we want userspace to be eventually consistent with
the state of the machine, but we don't want the overhead of immediately
waking the agent to deal with the message.

Change-Id: I10d81c174b3a2da8da0517f5bf0037c51d4dae0c