sched: cleanup ghost run_flag DO_NOT_PREEMPT
Currently DO_NOT_PREEMPT only applies to ghost tasks. This patch expands
that so that we really won't preempt any non-idle task, which is
slightly more useful.

Additionally, DO_NOT_PREEMPT should ignore the case where we are
committing the same task that is currently running, which can happen
often with sync commits.

Nothing currently uses this flag, so no userspace code needs to be
changed.

Change-Id: Ib7991b9c63d7d51467aea47a2b7c9f1c2781f3a2