sched: wake agent in release_from_ghost before losing the status_word
release_from_ghost() is slightly special because towards the end of it
we set p->ghost.status_word to NULL because the kernel is not supposed
to access it anymore (we have transferred ownership to userspace by
setting GHOST_SW_F_CANFREE).

Prior to this change we produced a MSG_TASK_DEPARTED or MSG_TASK_DEAD
and called ghost_wake_agent_of(p) only _after_ the status_word was
made inaccessible to the kernel. This is not required and hinders
wakeup_cpu preference logic in upcoming commits.

Fix this by waking the agent immediately after producing the msg
so that the wakeup preference selection logic has full access to
the status_word.

Change-Id: I4baee86bac3adf8342a64195e15c6530aaaf5572