sched: elide ghost_commit_greedy_txn() if ghost is not active.
Prior to this change the static key check for !ghost_active was
done inside the ghost_commit_greedy_txn() function. On hosts that
are don't enable ghost this is effectively a 'call' immediately
followed by a 'ret'.

However in certain syscall intensive benchmarks even this was
visible overhead in perf profiles.

Fix this by checking the 'ghost_active' static_key prior to
making the function call.

Change-Id: I30452aa9096d1a2a912cb243b8c11b271a5b9d25